Extract UserDetail in its own component
